requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.17.0
requests>=2.31.0
torchvision
